+++ src/domain/mod.rs
@@ -9,6 +9,7 @@ mod synapse;
 pub use activation::Activation;
 pub use events::{
     Event, RandomNeuronAdded, RandomNeuronRemoved, RandomSynapseAdded, RandomSynapseRemoved,
+    SynapseWeightMutated,
 };
 pub use network::Network;
 pub use neuron::Neuron;